requirements.txt
streamlit==1.32.0
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langchain-iris==0.0.1
intersystems-iris==1.0.0
openai==1.12.0
python-dotenv==1.0.1
plotly==5.18.0
pandas==2.1.4
